Runtime Optimization: Process all actions possible within 30ms frame rate
Runtime will keep processing actions that do not produce output until the frame timeout of 30ms.
This is an abitrary value and (TODO) we should connect it with the Choreographer frame clock
in similar way to how molecule does it with the AndroidUiDispatcher.Main and its frame clock.

Tests are modified for when using the optimized runtime in order to wait past the timeout.

We add expect fun currentTimeMillis() to the common code to get epoch millis on all platforms.